Restructured project for maven
