+++ src/main/java/ua/lviv/iot/models/Averruncator.java
@@ -1,4 +1,4 @@
-package ua.lviv.iot.lab_3.models;
+package ua.lviv.iot.models;
 
 public class Averruncator extends GardenTool {
     private Hand hand;
