+++ src/main/java/ua/lviv/iot/models/GardenTool.java
@@ -1,4 +1,4 @@
-package ua.lviv.iot.lab_3.models;
+package ua.lviv.iot.models;
 
 public abstract class GardenTool {
     private double weight;
